+++ client/js/controllers/results-controller.js
@@ -221,24 +221,27 @@ $scope.updateProject = function() {
 $scope.deleteResult = function(result, identifier) {
   var i = $scope.project._id;
   var id = identifier;
-  $http.delete('/api/projects/' + i + '/' + id)
-    .success(function() {
-      console.log("success");
-      var idx = $scope.results.indexOf(result);
-      if (idx >= 0) {
-        $scope.results.splice(idx, 1);
-      }
-      $scope.updateProject();
-      refresh();
-    })
-    .error(function() {
-      var idx = $scope.results.indexOf(result);
-      if (idx >= 0) {
-        $scope.results.splice(idx, 1);
-      }
-      $scope.updateProject();
-      refresh();
-    });
+  var r = confirm("Are you sure you want to delete the results "+result.name+ "?");
+  if(r){
+    $http.delete('/api/projects/' + i + '/' + id)
+      .success(function() {
+        console.log("success");
+        var idx = $scope.results.indexOf(result);
+        if (idx >= 0) {
+          $scope.results.splice(idx, 1);
+        }
+        $scope.updateProject();
+        refresh();
+      })
+      .error(function() {
+        var idx = $scope.results.indexOf(result);
+        if (idx >= 0) {
+          $scope.results.splice(idx, 1);
+        }
+        $scope.updateProject();
+        refresh();
+      });
+  }
 }
 
 //Reload the data from the result into the current data of the project
@@ -682,6 +685,8 @@ app.directive('exportEverythingToCsv',function($timeout, $http){
         var el = element[0];
         element.bind('click', function(e){
           var zip = new JSZip();
+          document.getElementById('exportRS').style.display = "none";
+          document.getElementById('exportRF').style.display = "none";
           if(document.getElementById('res'+scope.values).checked == true){
             var table = document.getElementById("resultsTable"+scope.values);
             var csvString = '';
@@ -788,11 +793,14 @@ app.directive('exportEverythingToCsv',function($timeout, $http){
             zip.file(scope.names+"/notes.txt", csvString);
           }
           if(document.getElementById('res'+scope.values).checked == true || document.getElementById('cri'+scope.values).checked == true || document.getElementById('alt'+scope.values).checked == true || document.getElementById('per'+scope.values).checked == true || document.getElementById('cat'+scope.values).checked == true || document.getElementById('par'+scope.values).checked == true || document.getElementById('pro'+scope.values).checked == true || document.getElementById('note'+scope.values).checked == true){
+            document.getElementById('exportRS').style.display = "";
             zip.generateAsync({type:"blob"})
             .then(function(content) {
                 // see FileSaver.js
                 saveAs(content, scope.names+".zip");
             });
+          }else{
+            document.getElementById('exportRF').style.display = "";
           }
         });
       }, 1000);
